#7.4 Protecting Video Routes

+++ controllers/videoController.js
@@ -54,7 +54,6 @@ export const videoDetail = async (req, res) => {
 
   try {
     const video = await Video.findById(id).populate("creator");
-    console.log(video);
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -68,6 +67,10 @@ export const getEditVideo = async (req, res) => {
 
   try {
     const video = await Video.findById(id);
+    if (video.creator !== req.use.id) {
+      throw Error();
+    }
+
     res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -94,10 +97,13 @@ export const deleteVideo = async (req, res) => {
   } = req;
 
   try {
-    await Video.findByIdAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.use.id) {
+      throw Error();
+    } else {
+      await Video.findByIdAndRemove({ _id: id });
+    }
   } catch (error) {
-    console.log(error);
+    res.redirect(routes.home);
   }
-
-  res.redirect(routes.home);
 };
